Move some set parse_date_cols to pythong parser
